factor out counting dict
